requirements.txt
llama-index
pypdf
transformers
torchvision
python-telegram-bot
python-dotenv
